The Mixing Matrix of a Continuous-Time Quantum Walk

The Mixing Matrix of a Continuous-Time Quantum Walk

Usage

# S3 method for ctqwalk
mixing_matrix(object, t, ...)

Arguments

object

an instance of class ctqwalk.

t

it will be returned the mixing matrix at time t.

...

further arguments passed to or from other methods.

Value

mixing_matrix() returns the mixing matrix of the CTQW evaluated at time t.

Details

Let \(U(t)\) be the time evolution operator of the quantum walk at time \(t\), then the mixing matrix is given by

$$M(t) = U(t) \circ \overline{U(t)}$$

\(M(t)\) is a doubly stochastic real symmetric matrix, which encodes the probability density of the quantum system at time \(t\).

More precisely, the \((M(t))_{ab}\) entry gives us the probability of measuring the standard basis state \(|b \rangle\) at time \(t\), given that the quantum walk started at \(|a \rangle\).

See also

Examples

walk <- ctqwalk(matrix(c(0,1,0,1,0,1,0,1,0), nrow=3))

# Returns the mixing matrix at time t = 2*pi, M(2pi)
mixing_matrix(walk, t = 2*pi)
#>             [,1]      [,2]        [,3]
#> [1,] 0.005025663 0.1317325 0.863241848
#> [2,] 0.131732489 0.7365350 0.131732489
#> [3,] 0.863241848 0.1317325 0.005025663